requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.3
Flask-WTF==1.1.1
WTForms==3.0.1
requests==2.31.0
openai==1.3.0
google-cloud-texttospeech==2.16.3
python-dotenv==1.0.0
APScheduler==3.10.4
redis==5.0.1
gunicorn==21.2.0
Werkzeug==2.3.7
SQLAlchemy==2.0.23
alembic==1.12.1
python-multipart==0.0.6
celery==5.3.4 